backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.16.4",
    "bcrypt==3.2.0",
    "fastapi>=0.116.1",
    "jose>=1.0.0",
    "jwt>=1.4.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.10",
    "pydantic[email]>=2.11.7",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
    "websockets>=15.0.1",
]
